moved GameStartupEvent in correct package
